Addressed Simon's Comments Pt. 1
