+++ slsim/Microlensing/lightcurve.py
@@ -14,10 +14,8 @@ from slsim.Util.astro_util import (
     extract_light_curve,
 )
 
-from slsim.Microlensing.source_morphology import (
-    GaussianSourceMorphology,
-    AGNSourceMorphology,
-)
+from slsim.Microlensing.source_morphology.agn import AGNSourceMorphology
+from slsim.Microlensing.source_morphology.gaussian import GaussianSourceMorphology
 
 
 class MicrolensingLightCurve(object):
@@ -33,7 +31,7 @@ class MicrolensingLightCurve(object):
     ):
         """
         :param magnification_map: MagnificationMap object, if not provided.
-        :param time_duration: Time duration for which the lightcurve is needed (in days).
+        :param time_duration: Time duration for which the lightcurve is needed (in days). In observer frame (z = 0).
         :param point_source_morphology: Type of source morphology to use. Default is 'gaussian'. Options are 'gaussian' or 'agn' (Accretion Disk) or 'supernovae'.
         :param kwargs_source_morphology: Dictionary of keyword arguments for the source morphology class. This should be as per the source morphology type.
 
@@ -47,49 +45,58 @@ class MicrolensingLightCurve(object):
             "eddington_ratio": eddington_ratio, }.
         """
 
-        self.magnification_map = magnification_map
-        self.time_duration = time_duration
-        self.point_source_morphology = point_source_morphology
-        self.kwargs_source_morphology = kwargs_source_morphology
+        self._magnification_map = magnification_map
+        self._time_duration_observer_frame = time_duration
+
+        self._point_source_morphology = point_source_morphology
+        self._kwargs_source_morphology = kwargs_source_morphology
 
         # Initialize the convolved map and source morphology
-        self.convolved_map = None
-        self.source_morphology = None
+        self._convolved_map = None
+        self._source_morphology = None
 
     def get_convolved_map(
         self,
         return_source_morphology=False,
     ):
-        """Get the convolved map based on the source morphology."""
-        if self.point_source_morphology == "gaussian":
+        """Get the convolved map based on the source morphology.
+        
+        :param return_source_morphology: Whether to return the source
+            morphology object or not. Default is False.
+        
+        :return: The convolved map and the source morphology object if
+            requested. Otherwise, only the convolved map is returned.
+        :rtype: numpy.ndarray or tuple
+        """
+        if self._point_source_morphology == "gaussian":
             # Gaussian source morphology
             source_morphology = GaussianSourceMorphology(
-                **self.kwargs_source_morphology,  # sets the source size, redshift, and cosmology
-                length_x=self.magnification_map.half_length_x * 2,
-                length_y=self.magnification_map.half_length_y * 2,
-                num_pix_x=self.magnification_map.num_pixels_x,
-                num_pix_y=self.magnification_map.num_pixels_y,
+                **self._kwargs_source_morphology,  # sets the source size, redshift, and cosmology
+                length_x=self._magnification_map.half_length_x * 2,
+                length_y=self._magnification_map.half_length_y * 2,
+                num_pix_x=self._magnification_map.num_pixels_x,
+                num_pix_y=self._magnification_map.num_pixels_y,
                 center_x=0,
                 center_y=0,
             )
 
             # convolve the magnification map with the Gaussian kernel
-            self.convolved_map = fftconvolve(
-                self.magnification_map.magnifications,
+            self._convolved_map = fftconvolve(
+                self._magnification_map.magnifications,
                 source_morphology.kernel_map,
                 mode="same",
             )
 
-        elif self.point_source_morphology == "agn":
+        elif self._point_source_morphology == "agn":
             # AGN source morphology
             source_morphology = AGNSourceMorphology(
-                **self.kwargs_source_morphology,
+                **self._kwargs_source_morphology,
             )
             cosmo = source_morphology.cosmo
             source_redshift = source_morphology.source_redshift
 
             # magnification map pixel size
-            pixel_size_magnification_map = self.magnification_map.get_pixel_size_meters(
+            pixel_size_magnification_map = self._magnification_map.get_pixel_size_meters(
                 source_redshift=source_redshift, cosmo=cosmo
             )
 
@@ -107,11 +114,11 @@ class MicrolensingLightCurve(object):
             rescaled_kernel_map = rescaled_kernel_map / np.nansum(rescaled_kernel_map)
 
             # convolve the magnification map with the kernel map, #TODO: make this a cross-correlation
-            self.convolved_map = fftconvolve(
-                self.magnification_map.magnifications, rescaled_kernel_map, mode="same"
+            self._convolved_map = fftconvolve(
+                self._magnification_map.magnifications, rescaled_kernel_map, mode="same"
             )
 
-        elif self.point_source_morphology == "supernovae":
+        elif self._point_source_morphology == "supernovae":
             # Supernovae source morphology
             source_morphology = None
             raise NotImplementedError(
@@ -124,12 +131,12 @@ class MicrolensingLightCurve(object):
             )
 
         if source_morphology:
-            self.source_morphology = source_morphology
+            self._source_morphology = source_morphology
 
         if return_source_morphology:
-            return self.convolved_map, source_morphology
+            return self._convolved_map, source_morphology
         else:
-            return self.convolved_map
+            return self._convolved_map
 
     def generate_lightcurves(
         self,
@@ -141,13 +148,11 @@ class MicrolensingLightCurve(object):
         x_start_position=None,
         y_start_position=None,
         phi_travel_direction=None,
-        return_track_coords=False,
-        return_time_array=False,
     ):
         """Generate lightcurves for a point source based on the convolved map.
 
         :param source_redshift: Redshift of the source
-        :param cosmo: Cosmology object for the lens class
+        :param cosmo: astropy.cosmology instance for the lens class
         :param lightcurve_type: Type of lightcurve to generate, either
             'magnitude' or 'magnification'. If 'magnitude', the
             lightcurve is returned in magnitudes normalized to the macro
@@ -158,19 +163,19 @@ class MicrolensingLightCurve(object):
             source plane (in km/s)
         :param num_lightcurves: Number of lightcurves to generate.
             Default is 1.
-        :param x_start_position: Starting x position of the lightcurve
-            in pixel coordinates. Default is None.
-        :param y_start_position: Starting y position of the lightcurve
-            in pixel coordinates. Default is None.
+        :param x_start_position: Starting x position of the lightcurve on the magnification map in arcsec. A value of 0 indicates the center of the magnification map. Default is None. If None, a random position is chosen.
+        :param y_start_position: Starting y position of the lightcurve on the magnification map in arcsec. A value of 0 indicates the center of the magnification map. Default is None. If None, a random position is chosen.
         :param phi_travel_direction: Angle of the travel direction in
-            degrees. Default is None.
-        :param return_track_coords: Whether to return the track
-            coordinates of the lightcuve(s) or not. Default is False.
-        :param return_time_array: Whether to return the time array used
-            for the lightcurve(s) or not. Default is False.
-        :return: A tuple of lightcurves, tracks, and time arrays if
-            requested. If only lightcurves are requested, a list of
-            lightcurves is returned.
+            degrees. Default is None. If None, a random angle is chosen. A value of 0
+            implies the positive x-axis of the magnification map.
+        :return: A tuple of lightcurves, tracks, and time arrays.
+
+            lightcurves: list of lightcurves
+            
+            tracks: x and y positions (in pixels) on the magnification map grid for the paths used to generate the lightcurves.
+
+            time_arrays: list of time arrays for each lightcurve
+
         """
 
         # Get the convolved magmap after convolving the magnification map with a Gaussian kernel
@@ -179,11 +184,29 @@ class MicrolensingLightCurve(object):
         )
 
         # determine physical pixel sizes in source plane
-        pixel_size_magnification_map = self.magnification_map.get_pixel_size_meters(
+        pixel_size_magnification_map = self._magnification_map.get_pixel_size_meters(
             source_redshift=source_redshift, cosmo=cosmo
         )
 
+        # convert x and y positions from arcsec to pixel coordinates on the magnification map grid
+        if x_start_position is not None:
+            x_start_position = (
+                x_start_position / self._magnification_map.half_length_x
+            ) * self._magnification_map.num_pixels_x / 2
+            x_start_position = int(
+                x_start_position + self._magnification_map.num_pixels_x // 2
+                )
+        
+        if y_start_position is not None:
+            y_start_position = (
+                y_start_position / self._magnification_map.half_length_y
+            ) * self._magnification_map.num_pixels_y / 2
+            y_start_position = int(
+                y_start_position + self._magnification_map.num_pixels_y // 2
+                )
+
         return self._generate_lightcurves(
+            source_redshift=source_redshift,
             convolved_map=convolved_map,
             pixel_size_magnification_map=pixel_size_magnification_map,
             num_lightcurves=num_lightcurves,
@@ -192,12 +215,11 @@ class MicrolensingLightCurve(object):
             x_start_position=x_start_position,
             y_start_position=y_start_position,
             phi_travel_direction=phi_travel_direction,
-            return_track_coords=return_track_coords,
-            return_time_array=return_time_array,
         )
 
     def _generate_lightcurves(
         self,
+        source_redshift,
         convolved_map,
         pixel_size_magnification_map,
         num_lightcurves=1,
@@ -206,11 +228,10 @@ class MicrolensingLightCurve(object):
         x_start_position=None,
         y_start_position=None,
         phi_travel_direction=None,
-        return_track_coords=False,
-        return_time_array=False,
     ):
         """Generate lightcurves for a point source based on the convolved map.
 
+        :param source_redshift: Redshift of the source
         :param convolved_map: Convolved magnification map
         :param pixel_size_magnification_map: Pixel size of the
             magnification map in meters
@@ -225,26 +246,33 @@ class MicrolensingLightCurve(object):
         :param effective_transverse_velocity: Transverse velocity in
             source plane (in km/s)
         :param x_start_position: Starting x position of the lightcurve
-            in pixel coordinates. Default is None.
+            in pixel coordinates. Default is None. If None, a random
+            position is chosen. A value of 0 indicates the center of the
+            magnification map.
         :param y_start_position: Starting y position of the lightcurve
-            in pixel coordinates. Default is None.
+            in pixel coordinates. Default is None. If None, a random
+            position is chosen. A value of 0 indicates the center of the
+            magnification map.
         :param phi_travel_direction: Angle of the travel direction in
-            degrees. Default is None.
-        :param return_track_coords: Whether to return the track
-            coordinates of the lightcuve(s) or not. Default is False.
-        :param return_time_array: Whether to return the time array used
-            for the lightcurve(s) or not. Default is False.
+            degrees. Default is None. If None, a random angle is chosen. A value of 0
+            implies the positive x-axis of the magnification map.
         :return: A tuple of lightcurves, tracks, and time arrays if
-            requested. If only lightcurves are requested, a list of
-            lightcurves is returned.
+            requested.
         """
 
         LCs = []
         tracks = []
         time_arrays = []
 
+
+        # time duration in source frame
+        self._time_duration_source_frame = (
+            self._time_duration_observer_frame
+            / (1 + source_redshift)
+        )
+
         time_duration_years = (
-            self.time_duration / 365.25
+            self._time_duration_source_frame / 365.25
         )  # converting time_duration from days to years
 
         for _ in range(num_lightcurves):
@@ -264,7 +292,7 @@ class MicrolensingLightCurve(object):
             if lightcurve_type == "magnitude":
                 # print("Extracting magnitude for light curve...")
                 light_curve = -2.5 * np.log10(
-                    light_curve / np.abs(self.magnification_map.mu_ave)
+                    light_curve / np.abs(self._magnification_map.mu_ave)
                 )
             elif lightcurve_type == "magnification":
                 # print("Extracting magnification for light curve...")
@@ -275,110 +303,6 @@ class MicrolensingLightCurve(object):
                 )
             LCs.append(light_curve)
             tracks.append(np.array([x_positions, y_positions]))
-            time_arrays.append(np.linspace(0, self.time_duration, len(light_curve)))
+            time_arrays.append(np.linspace(0, self._time_duration_observer_frame, len(light_curve)))
 
-        if return_track_coords and not (return_time_array):
-            return LCs, tracks
-
-        if return_time_array and not (return_track_coords):
-            return LCs, time_arrays
-
-        if return_track_coords and return_time_array:
-            return LCs, tracks, time_arrays
-
-        if not (return_track_coords) and not (return_time_array):
-            return LCs
-
-    def plot_lightcurves_and_magmap(
-        self, lightcurves, tracks=None, lightcurve_type="magnitude"
-    ):
-        """Plot the point source lightcurve."""
-        fig, ax = plt.subplots(1, 2, figsize=(18, 6), width_ratios=[2, 1])
-
-        time_array = np.linspace(0, self.time_duration, len(lightcurves[0]))  # in days
-
-        # light curves
-        for i in range(len(lightcurves)):
-            ax[0].plot(time_array, lightcurves[i], label=f"Lightcurve {i+1}")
-        ax[0].set_xlabel("Time (days)")
-
-        if lightcurve_type == "magnitude":
-            ax[0].set_ylabel(
-                "Magnitude $\\Delta m = -2.5 \\log_{10} (\\mu / \\mu_{\\text{av}})$"
-            )
-            im_to_show = -2.5 * np.log10(
-                self.convolved_map / np.abs(self.magnification_map.mu_ave)
-            )
-        elif lightcurve_type == "magnification":
-            ax[0].set_ylabel("Magnification $\\mu$")
-            im_to_show = self.convolved_map
-
-        ax[0].set_ylim(np.nanmin(im_to_show), np.nanmax(im_to_show))
-        ax[0].legend()
-
-        # magmap
-        conts = ax[1].imshow(
-            im_to_show,
-            cmap="viridis_r",
-            extent=[
-                (self.magnification_map.center_x - self.magnification_map.half_length_x)
-                / self.magnification_map.theta_star,
-                (self.magnification_map.center_x + self.magnification_map.half_length_x)
-                / self.magnification_map.theta_star,
-                (self.magnification_map.center_y - self.magnification_map.half_length_y)
-                / self.magnification_map.theta_star,
-                (self.magnification_map.center_y + self.magnification_map.half_length_y)
-                / self.magnification_map.theta_star,
-            ],
-            origin="lower",
-        )
-        divider = make_axes_locatable(ax[1])
-        cax = divider.append_axes("right", size="5%", pad=0.05)
-        cbar = plt.colorbar(conts, cax=cax)
-        if lightcurve_type == "magnitude":
-            cbar.set_label(
-                "Microlensing $\\Delta m = -2.5 \\log_{10} (\\mu / \\mu_{\\text{av}})$ (magnitudes)"
-            )
-        elif lightcurve_type == "magnification":
-            cbar.set_label("Microlensing magnification $\\mu$")
-        ax[1].set_xlabel("$x / \\theta_★$")
-        ax[1].set_ylabel("$y / \\theta_★$")
-        # tracks are in pixel coordinates
-        # to map them to the magmap coordinates, we need to convert them to the physical coordinates
-        delta_x = (
-            2
-            * self.magnification_map.half_length_x
-            / self.magnification_map.num_pixels_x
-        )
-        delta_y = (
-            2
-            * self.magnification_map.half_length_y
-            / self.magnification_map.num_pixels_y
-        )
-        mid_x_pixel = self.magnification_map.num_pixels_x // 2
-        mid_y_pixel = self.magnification_map.num_pixels_y // 2
-        if tracks is not None:
-            for j in range(len(tracks)):
-                ax[1].plot(
-                    (tracks[j][1] - mid_x_pixel)
-                    * delta_x
-                    / self.magnification_map.theta_star,
-                    (tracks[j][0] - mid_y_pixel)
-                    * delta_y
-                    / self.magnification_map.theta_star,
-                    "w-",
-                    lw=1,
-                )
-                ax[1].text(
-                    (tracks[j][1][0] - mid_x_pixel)
-                    * delta_x
-                    / self.magnification_map.theta_star,
-                    (tracks[j][0][0] - mid_y_pixel)
-                    * delta_y
-                    / self.magnification_map.theta_star,
-                    str(j + 1),
-                    color="white",
-                    fontsize=16,
-                )
-
-        return ax
+        return LCs, tracks, time_arrays
