+++ slsim/Microlensing/source_morphology/supernovae.py
@@ -1,0 +1,16 @@
+__author__ = "Paras Sharma"
+
+from slsim.Microlensing.source_morphology.source_morphology import (
+    SourceMorphology,
+)
+
+
+class SupernovaeSourceMorphology(SourceMorphology):
+    """Class for Supernovae source morphology."""
+
+    def __init__(self, *args, **kwargs):
+        """Initializes the Supernovae source morphology."""
+        super().__init__(*args, **kwargs)
+        raise NotImplementedError(
+            "Supernovae source morphology is not implemented yet."
+        )
